+++ crates/builder/core/src/flashblocks/mod.rs
@@ -1,7 +1,7 @@
 //! Flashblocks builder types.
 
 mod best_txs;
-pub use best_txs::{BestFlashblocksTxs, RejectionCache};
+pub use best_txs::BestFlashblocksTxs;
 
 mod generator;
 pub use generator::{
@@ -21,7 +21,6 @@ pub use context::{
 };
 
 mod payload;
-pub use payload::FlashblocksExecutionInfo;
 
 mod service;
 pub use service::FlashblocksServiceBuilder;
